requirements.txt
streamlit
PyPDF2
numpy
faiss-cpu  # or any other vector store library you choose
transformers  # for embedding model